Add config for workers
Notice that we're referencing an ENV var here - this is a pattern
@ashkan18 turned me onto. The idea is that when it's undefined, then you
have 5 workers, but should you want to adjust this, you can set that var
and the Heroku dyno will have to be restarted, but not rebuild like it
would if you had to do a code change. Neato!