Once the basics are specified, check gem order
I like it when a Gemfile has the Rails version, db and web server right
at the top of the file - I find that easy to glance at and know what
sort of project I'm working with. Other than that, I'm totally down with
the alpha ordering.

Normally I would simply disable this cop, but I'm trying this approach
instead.