Don't specify production database
This file is actually written by Heroku when you deploy, so there's no
need for us to specify any of this stuff. What's more, if you do specify
this stuff, then when you run tasks like:

```
$ rake db:create:all
```

It'll try to create a production database when all you probably wanted
was development and test.