Switch to sql structure file
I prefer this approach to dumping the database schema - docs here:

http://edgeguides.rubyonrails.org/active_record_migrations.html#schema-dumping-and-you

I also ignore the file because changes to it are unimportant. The file
is automatically generated and different versions of Postgres can cause
slight variations in the order of the file, but it doesn't matter.

And if I were to get a little philosophical, I'd also say that I prefer
the SQL format because it forces us to read, well, SQL. When it comes to
looking at your database schema, I think less abstraction is better.

Finally, the ruby schema is not capable of expressing everything that
raw SQL can - for most cases this is fine, but should we need that sort
of thing, we'll have it without having to do anything.