Import from external URL
This was FAR harder than I expected it. What I thought would work was
something like this:

```ruby
data = Faraday.get(url).read
```

But I was wrong because encodings! What would happen is that the data
coming from Faraday would end up with an ASCII encoding that I couldn't
quite figure out - turns out that's the default encoding for `Net::HTTP`
and you have to work around it.

There's a Faraday middleware project that I could have used, but I
prefer the simplicity of this approach.