Don't automatically connect to external DBs
While setting up the project fresh, I was getting errors because I did
not have the ENV vars set for these service, but the vars were there.
That meant that when these files were evaluated, the connections would
fail and the application was not able to load.

My approach in this commit is to switch from a constant that connects
upon app boot to a method that returns a connection. This allows for
these files to be evaluated, but not bring down the app when the
configuration isn't set as expected.