Resolve organizations by source rank
This is the punchline - we add a rank column to the sources table and
then choose the organization name, location and website based on the
source of data with the best rank (lower is better).

This commit also separates the business of creating the CSV and
resolving the organization, thanks for the insight on that @gnilekaw!