Set transaction isolation level to prevent duplicates.
This allows the job to be moved back into async `perform_later`,
but there is a side effect. Now there are occassional errors for:
'ERROR: could not serialize access due to read/write dependencies
among transactions'. This points out that we need some error
handling for when this job fails.