Simplify relationship between orgs and tags
Early on in our understanding of the system, we decided that we needed a
model that joined imports, organizations and tags. As time has passed,
it has become clear that this was more complicated than we needed.

With this commit I've severed the connection to imports - now we simply
have a join model between organizations and tags.