Support tags in CSV import rake task
This commit adds PaperTrail to `OrganizationTag` and then updates the
collaborator classes to the import rake task so that they pay attention
to a column I'm calling `tag_names`. The idea here is to make this be a
comma-separated list of tag names.

The actual work of tagging, I've pushed down to the `OrganizationTag`
class. It seemed like a good place for the knowledge of what to do in
the various combinations that we could get ourselves into.

Note: I have disabled a RuboCop violation - you mad? I feel bad about
this, but not enough to refactor. I know that this part of the code is
going to have conflicts with #65 and once I get to #73 and #74, I will
be adding even more here.

I guess my point is that I'd like to feel the pain a little longer
before I figure out what I'd like to do here. And I reserve the right to
continue ignoring, haha! Unless someone feels strongly or has a better
idea, of course. :wink: