Add basic import integration tests
This was actually super challenging for me! One thing that was tricky
was figuring out how to assert about stdout, stderr and status codes.
Honeybadger to the rescue:

http://blog.honeybadger.io/capturing-stdout-stderr-from-shell-commands-via-ruby/

Then, later, I was really struggling with fabricating sources and
getting them in the context of the rake task. Turns out, I had to turn
off transactional fixtures, so I made a separate spec helper file for
that. Interested in thoughts on that for sure.

Finally, I'll point out that I've added a fixtures folder to the spec
dir where we can throw different csv files to import and then assert
about.