Use inline adapter for testing rake tasks
The goal of this commit is to improve our test coverage by actually
having the `RawInputTransformJob` run. In order to get these jobs to
run, you have to use the inline adapter, BUT you can't set this up in
the test process because the rake task is actually running a different
process! That left me with some bad choices, but I settled on the
approach here - just go ahead and pass in an ENV var that sets the
adapter and when the rake task loads it's environment file, it'll get
picked up there. Tricksy!

The punchline, then, is that I get to assert about the models that were
created as a result of the import taking place and now I have coverage
all up and down the stack of this rake task.