Simplify repetitive test assertions
I don't always like to refactor tests like this, but it was getting a
little silly and there was actually a bug! If you notice in the diff, we
had this:

```ruby
website = Location.first
```

This this refactoring, that type of bug wouldn't be possible and it'll
be easier to add any additional types of relationships here as well.