Add Sidekiq console
For now I'm just using the basic auth approach - if we end up with a
more fancy authentication strategy, then this could change.

Another thing I did here that's different than the docs is move the
settings out of the routes file and into an initializer. I think this is
cleaner, but I didn't leave a comment in case it's confusing at all.

closes #85