Add a more realistic test case for imports
As @gnilekaw rightly points out the two test cases here are probably not
very realistic - what's going to probably be typical is that imports
contain only a subset of the available fields. This commit just adds a
test for that use case to improve our confidence that the in-between
case is covered.