Surface more details on import errors
With this commit, @gnilekaw and I have once again refactored the heck
out of the `RawInputChanges` class. What we figured out was that in
order to give ourselves as much information as possible about why a
`RawInput` could not be applied, it was best to build all relations and
then try to save them. With that approach, we're actually able to get
all the errors and save them to the `RawInput` for further
investigation.