Add more granularity to source rankings
Prior to this commit rank was all or none - either one source was better
than another or not. What was asked for was a way for sources to be
ranked for each individual field for an organization.

The first thing that came to mind was to go from one column for rank to
one column per field for rank and that's what this commit does.