Refactor rank_for to accept symbol
We often want to metaprogram around the rank types for a source and
we've been settling for `public_send`, but that's gross! A better
approach is to rethink the `rank_for` method. Now we convert the
Rankable into a symbol that we can pass and look up in a mapping.

This is better than the previous implementation because it's more
flexible - any caller can pass a symbol and there's no longer this
coupling between the Rankable classes and the `rank_for` method.