Install micromachine and configure for imports
I looked for state machine libraries to use and found that there were
some *really* complex options. The most popular choice seems to be
`pluginaweek/state_machine` and after reading through it, I was
overwhelmed. All I really needed was something to keep track of states
and ensure I didn't make mistakes - there were SO many features that I
didn't care about.

When I found micromachine, I liked it a lot more - quite simple! The
entire library is one file and can easily be read.

Now, it might be that it won't be enough - that we'll end up wanting
more features, but I wanted to start small and only go there if I have
to.