Centralize knowledge about working with Imports
Rather than spreading the knowledge about how Imports go through their
various states, centralize this to the model. This *does* mean that the
model knows a little more than it did before, but I think that's for the
best - it's now harder to make a mistake because you're meant to go
through the model, so there's just one place to detail these jobs.