Ensure environment is setup for tests
Prior to this commit dotenv wasn't being loaded for the test
environment, which meant that basic auth tests were blowing up. I've
also included a test version with just the config required.

This will probably change somewhat when the new encrypted secrets stuff
lands in Rails 5.1, but this should get the build back to green.