Replace export task with data warehouse syncing
This is the punchline - this commit removes the rake task in favor of a
job that will export all Bearden data to S3 and then import it into
Redshift.

In order to do this, we had to have a job - that's a given. And the
Redshift module that handles the connections abstracts that away. What
was left to do was decide which how the rest of the work was going to
get done and who was doing what.

After pairing with @gnilekaw and then finishing this thing off, I think
I've settled on some nice abstractions. The only thing that sorta makes
me feel bad is the similarity in names between `DataWarehouse` and
`DataWarehouseReset`. Words just failed me when trying to tease apart
those two things.