Export in batches
@gnilekaw found that exporting the entire Organization table on staging
didn't work because we'd run out of memory. As an experiment, I moved to
doing this work in batches - we create a CSV for every 10,000 records
and then loop through it when copying over to Redshift.

This works in the sense that the DataWarehouseSyncJob now finishes
without brining the dynos down, but it's still slow and using more than
1GB of memory with only 45k records.

So, this commit closes #148, but it's not the long-term solution we're
looking for.