Convert import data to UTF-8 when possible
Rather than forcing the encoding, this commit uses charlock_holmes to
gracefully convert data to UTF-8. I have yet to find data where this
conversion wasn't right, but it's possible we could still have imports
that raise when parsing the CSV.

But, this is much better than we had before and should solve the
majority case.

closes #164