Install Webmock
As a part of this work, I took a step back and looked at what mocking we
were already doing in the test suite so that I could find places where
we could start using Webmock instead. Here are some places I researched:

* spec/jobs/parse_csv_import_job_spec.rb
* spec/features/csv_import_spec.rb
* SlackBot mock in spec/rails_helper.rb

In all three cases, I gave Webmock a shot, I really did, but ended up
not using it. After looking at what the code would look like after the
change, I decided that the current mocking approach was better - mocking
at the class level was actually better.

I was surprised that this was what I found - I assumed that after
putting Webmock in, that I'd also move a bunch of specs to using it.
Instead, I've formed an opinion that Webmock is usually an anti-pattern
and yet here I am with a commit that adds it.

The reason that I feel good about adding it is that it'll provide a few
things:

* it will ensure that the test suite does not make any network calls
that aren't dealt with
* it will allow me to write tests that *do* need to integrate over
Faraday

The latter won't come up until a later commit, but the former is
interesting - this is a case where I'm glad for the exception, but don't
like the suggestion. Basically, if you accidentally write a spec that
hits the network, then Webmock will raise and suggest a `stub_request`
line that would match it. This is a cool feature of the library!

Having said that, resist! Most of the time, the better answer is to stub
at the Faraday-level or maybe you have another class you can stub out.