Fix: ensure options have sensible values
Prior to this commit the values of the options that we were producing
were the strings themselves. Luckily they started with a number, so when
they were saved to Postgres things Just Worked :tm:, but that was bad. A
better approach is to return an nested array of label, value as DHH
intended.