Allow admins to edit source ranks
This ended up opening a can of worms on the resolution of source ranks.
What I had was too naive - it only considered the case where you were
adding a new Source. What I had to do was improve this code so that it
supported the case of promoting/demoting existing source ranks.

Luckily, I was able to convince the team to only support editing one
Source at a time, so that means that we have far more simple resolution
logic than we might have otherwise. Even still, this broke my brain for
quite a while!