Use better names when testing sources
@gnilekaw expressed some confusion about the names of the sources that I
used. His suggestion was to flip things around, but when I tried that,
it was still sorta confusing to me. Then I tried going in a different
direction and using letters to help distinguish between the various
sources and I think it turned out better. Especially in the
demote/promote tests, I think it's now very clear what the order was
before/after the resolving.