Filter out actors that are not RawInput records
Prior to this commit any actor that wasn't a RawInput would have
actually caused an exception to be raised when the Rankable was
resolved. What this commit does is filter the actors down to just those
that can return a source. In an effort to support this direction, I've
also added a delegate call on RawInput up to the Import's Source.