Add basic length validations
We ran into a case where we had data that was too long for Redshift and
that got us thinking about what Bearden's responsibility should be here.
We don't want to impose some artificial limit on what is a valid URL,
for example, but we should also be suspicious of a URL that's longer
than 256 characters.

So, that's where I've set the limit on this commit. As RawInput records
are applied, rows will be blown up if they contain info that exceeds
this limit. A human can then review that row and will mostly likely see
that we have some bad data.

If that's not the case - the human investigates and it's a legit URL,
for example, then we can have a discussion about what to do. Probably,
this won't happen, but it's possible.