Add temp task to migrate tags to types
This commit adds a temporary rake task that migrates some tags over to
types. Once this migration is complete, this task can be removed from
the project. I took inspiration for this from a [thoughtbot][tb] post.

[tb]: https://robots.thoughtbot.com/data-migrations-in-rails

Note that the actor of the OrganizationTag is carried over to the
OrganizationType so that we don't lose that signal. This is necessary to
ensure that when we resolve the type of an organization our Rankable
logic doesn't get goofed up. I note it because, technically, that
RawInput record did NOT create that OrganizationType, this rake task
did.

closes #228