Don't create types that already exist
This is nice because when I ran this on staging the dyno ran out of
memory and I went to run it again forgetting that types would be created
more than once. Now, you can run this task as many times as you like and
only types that haven't been migrated will be affected.