Use default artsy-auth forbidden behavior
It turns out that artsy-auth already will return a forbidden response if
`authorized_artsy_token?` returns false, so there's really no need to do
this in our application code:

https://github.com/artsy/artsy-auth/blob/master/lib/artsy-auth/authenticated.rb#L13

It's true that with this commit we're going from a 404 to a 401, but I
can live with that. If this were not an internal tool, I might care a
little more about the status code, but for something that Artsy admins
use, seems legit.