Fix broken nav
When I upgraded Bootstrap all the tests passed, but the navigation was
broken and I didn't ever actually spin up a server and check. :cry:

This commit fixes that by adjusting a couple classes. I also fixed a
mistake where we weren't highlighting syncs when that was the page you
were on.