Upgrade RuboCop
The biggest thing here is that RuboCop is now enforcing the dependent
option on associations. There really isn't a concept of deletion in
Bearden at the moment, so my solution was to set this option to
`restrict_with_exception`. My thinking here is that if I get into a
situation where I want to delete something that has associations, then
this exception will force me to make some choices. Basically, I'm
kicking this can down the road because today I don't have any need to
delete, but if Future Jon does, then he'll be in a better position to
decide which of these options is appropriate.