Reorganize search specs
With this commit I'm splitting up the main use case for search -
searching for an organization by it's name from secondary ones. Since
we're developing this search API before the client, it's possible that
we've done more than is required. As far as I knew, only name was
required. If that ends up being the case, then I can just nuke that one
context.

Also, this gives me a place from which to add more rigorous testing
around partial matches, etc.