Set a default number of results
Prior to this commit there was no guidance for the number of results to
return - now we're hardcoding a size of 20, but also allowing the client
of the endpoint to supply their own value for this.

This is a good solution for our current use case, but might not work for
future consumers of this endpoint. Once we have to support clients that
want to page through the results, then we can switch gears and implement
proper pagination with something like Kaminari.