Fix search endpoint
When I merged the GraphQL work to master it was deployed to staging, so
I went to go smoke test it there, but was getting this:

```ruby
NameError (uninitialized constant #<Class:0x007fd422b59b10>::Search
```

So the `autoload_paths` wasn't working as expected. I tinkered with this
for a bit, but didn't make much progress. Instead, I'm just moving this
file into the models folder and removing the namespace.

Open to other ideas, but this should allow me to deploy this work to
staging and production.