Don't raise when the record can't be found
The return here only works if we use the `find_by` form, otherwise, find
will just raise if it can't find the record and you'll never even hit
the early return on the next line.

I know this because I was having issues with these jobs while testing
something locally here on my laptop. This change fixed those jobs from
failing, but might expose a different problem. I have a theory that
these search indexing jobs are enqueued and then worked too quickly for
Postgres. I think I was getting that exception about the record not
being found because it was a race condition with the database insert!

If that's true, then this "fix" would actually break our search indexing
and what we really need to do is find a way to introduce some delay
between insertion and indexing.